+++ src/api/webrpc/webrpc_test.go
@@ -103,7 +103,7 @@ func (fg *fakeGateway) InjectBroadcastTransaction(txn coin.Transaction) error {
 	return errors.New("fake gateway inject transaction failed")
 }
 
-func (fg fakeGateway) GetAddrUxOuts(addr cipher.Address) ([]*historydb.UxOutJSON, error) {
+func (fg fakeGateway) GetAddrUxOuts(addr []cipher.Address) ([]*historydb.UxOut, error) {
 	return nil, nil
 }