+++ src/wallet/service.go
@@ -24,29 +24,32 @@ type BalanceGetter interface {
 // Service wallet service struct
 type Service struct {
 	sync.RWMutex
-	wallets          Wallets
-	firstAddrIDMap   map[string]string // Key: first address in wallet; Value: wallet id
-	walletDirectory  string
-	cryptoType       CryptoType
-	disableWalletAPI bool
+	wallets         Wallets
+	firstAddrIDMap  map[string]string // Key: first address in wallet; Value: wallet id
+	walletDirectory string
+	cryptoType      CryptoType
+	enableWalletAPI bool
+	enableSeedAPI   bool
 }
 
 // Config wallet service config
 type Config struct {
-	WalletDir        string
-	CryptoType       CryptoType
-	DisableWalletAPI bool
+	WalletDir       string
+	CryptoType      CryptoType
+	EnableWalletAPI bool
+	EnableSeedAPI   bool
 }
 
 // NewService new wallet service
 func NewService(c Config) (*Service, error) {
 	serv := &Service{
-		firstAddrIDMap:   make(map[string]string),
-		cryptoType:       c.CryptoType,
-		disableWalletAPI: c.DisableWalletAPI,
+		firstAddrIDMap:  make(map[string]string),
+		cryptoType:      c.CryptoType,
+		enableWalletAPI: c.EnableWalletAPI,
+		enableSeedAPI:   c.EnableSeedAPI,
 	}
 
-	if serv.disableWalletAPI {
+	if !serv.enableWalletAPI {
 		return serv, nil
 	}
 
@@ -97,7 +100,7 @@ func NewService(c Config) (*Service, error) {
 func (serv *Service) CreateWallet(wltName string, options Options) (*Wallet, error) {
 	serv.Lock()
 	defer serv.Unlock()
-	if serv.disableWalletAPI {
+	if !serv.enableWalletAPI {
 		return nil, ErrWalletAPIDisabled
 	}
 	if wltName == "" {
@@ -113,6 +116,10 @@ func (serv *Service) CreateWallet(wltName string, options Options) (*Wallet, err
 func (serv *Service) ScanAheadWalletAddresses(wltName string, password []byte, scanN uint64, bg BalanceGetter) (*Wallet, error) {
 	serv.Lock()
 	defer serv.Unlock()
+
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
 
 	w, err := serv.getWallet(wltName)
 	if err != nil {
@@ -213,60 +220,68 @@ func (serv *Service) generateUniqueWalletFilename() string {
 }
 
 // EncryptWallet encrypts wallet with password
-func (serv *Service) EncryptWallet(wltID string, password []byte) error {
+func (serv *Service) EncryptWallet(wltID string, password []byte) (*Wallet, error) {
 	serv.Lock()
 	defer serv.Unlock()
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
+
 	w, err := serv.getWallet(wltID)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	if w.IsEncrypted() {
-		return ErrWalletEncrypted
+		return nil, ErrWalletEncrypted
 	}
 
 	if err := w.lock(password, serv.cryptoType); err != nil {
-		return err
+		return nil, err
 	}
 
 	// Save to disk first
 	if err := w.Save(serv.walletDirectory); err != nil {
-		return err
+		return nil, err
 	}
 
 	// Sets the encrypted wallet
 	serv.wallets.set(w)
-	return nil
+	return w, nil
 }
 
 // DecryptWallet decrypts wallet with password
-func (serv *Service) DecryptWallet(wltID string, password []byte) error {
+func (serv *Service) DecryptWallet(wltID string, password []byte) (*Wallet, error) {
 	serv.Lock()
 	defer serv.Unlock()
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
+
 	w, err := serv.getWallet(wltID)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	// Returns error if wallet is not encrypted
 	if !w.IsEncrypted() {
-		return ErrWalletNotEncrypted
+		return nil, ErrWalletNotEncrypted
 	}
 
 	// Unlocks the wallet
 	unlockWlt, err := w.unlock(password)
 	if err != nil {
-		return err
+		return nil, err
 	}
 
 	// Updates the wallet file
 	if err := unlockWlt.Save(serv.walletDirectory); err != nil {
-		return err
+		return nil, err
 	}
 
 	// Sets the decrypted wallet in memory
 	serv.wallets.set(unlockWlt)
-	return nil
+	return unlockWlt, nil
 }
 
 // NewAddresses generate address entries in given wallet,
@@ -275,6 +290,10 @@ func (serv *Service) DecryptWallet(wltID string, password []byte) error {
 func (serv *Service) NewAddresses(wltID string, password []byte, num uint64) ([]cipher.Address, error) {
 	serv.Lock()
 	defer serv.Unlock()
+
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
 
 	w, err := serv.getWallet(wltID)
 	if err != nil {
@@ -312,6 +331,10 @@ func (serv *Service) NewAddresses(wltID string, password []byte, num uint64) ([]
 func (serv *Service) GetAddresses(wltID string) ([]cipher.Address, error) {
 	serv.RLock()
 	defer serv.RUnlock()
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
+
 	w, err := serv.getWallet(wltID)
 	if err != nil {
 		return nil, err
@@ -324,6 +347,9 @@ func (serv *Service) GetAddresses(wltID string) ([]cipher.Address, error) {
 func (serv *Service) GetWallet(wltID string) (*Wallet, error) {
 	serv.RLock()
 	defer serv.RUnlock()
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
 
 	return serv.getWallet(wltID)
 }
@@ -338,21 +364,25 @@ func (serv *Service) getWallet(wltID string) (*Wallet, error) {
 }
 
 // GetWallets returns all wallet clones
-func (serv *Service) GetWallets() Wallets {
+func (serv *Service) GetWallets() (Wallets, error) {
 	serv.RLock()
 	defer serv.RUnlock()
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
+
 	wlts := make(Wallets, len(serv.wallets))
 	for k, w := range serv.wallets {
 		wlts[k] = w.clone()
 	}
-	return wlts
+	return wlts, nil
 }
 
 // ReloadWallets reload wallets
 func (serv *Service) ReloadWallets() error {
 	serv.Lock()
 	defer serv.Unlock()
-	if serv.disableWalletAPI {
+	if !serv.enableWalletAPI {
 		return ErrWalletAPIDisabled
 	}
 	wallets, err := LoadWallets(serv.walletDirectory)
@@ -371,6 +401,10 @@ func (serv *Service) CreateAndSignTransaction(wltID string, password []byte, vld
 	headTime, coins uint64, dest cipher.Address) (*coin.Transaction, error) {
 	serv.RLock()
 	defer serv.RUnlock()
+	if !serv.enableWalletAPI {
+		return nil, ErrWalletAPIDisabled
+	}
+
 	w, err := serv.getWallet(wltID)
 	if err != nil {
 		return nil, err
@@ -537,6 +571,10 @@ func (serv *Service) CreateAndSignAdvancedTransaction(advancedSpend AdvancedSpen
 func (serv *Service) UpdateWalletLabel(wltID, label string) error {
 	serv.Lock()
 	defer serv.Unlock()
+	if !serv.enableWalletAPI {
+		return ErrWalletAPIDisabled
+	}
+
 	var wlt *Wallet
 	if err := serv.wallets.update(wltID, func(w *Wallet) error {
 		w.setLabel(label)
@@ -550,10 +588,15 @@ func (serv *Service) UpdateWalletLabel(wltID, label string) error {
 }
 
 // Remove removes wallet of given wallet id from the service
-func (serv *Service) Remove(wltID string) {
+func (serv *Service) Remove(wltID string) error {
 	serv.Lock()
 	defer serv.Unlock()
+	if !serv.enableWalletAPI {
+		return ErrWalletAPIDisabled
+	}
+
 	serv.wallets.remove(wltID)
+	return nil
 }
 
 func (serv *Service) removeDup(wlts Wallets) Wallets {
@@ -595,4 +638,37 @@ func (serv *Service) removeDup(wlts Wallets) Wallets {
 	}
 
 	return wlts
+}
+
+// GetWalletSeed returns seed of encrypted wallet of given wallet id
+// Returns ErrWalletNotEncrypted if it's not encrypted
+func (serv *Service) GetWalletSeed(wltID string, password []byte) (string, error) {
+	serv.RLock()
+	defer serv.RUnlock()
+	if !serv.enableWalletAPI {
+		return "", ErrWalletAPIDisabled
+	}
+
+	if !serv.enableSeedAPI {
+		return "", ErrSeedAPIDisabled
+	}
+
+	w, err := serv.getWallet(wltID)
+	if err != nil {
+		return "", err
+	}
+
+	if !w.IsEncrypted() {
+		return "", ErrWalletNotEncrypted
+	}
+
+	var seed string
+	if err := w.guardView(password, func(wlt *Wallet) error {
+		seed = wlt.seed()
+		return nil
+	}); err != nil {
+		return "", err
+	}
+
+	return seed, nil
 }